requirements.txt
Flask==1.1.1
line-bot-sdk==1.12.1
psycopg2==2.7.5
SQLAlchemy==1.3.6